# public/openapi.yaml
openapi: 3.1.0
info:
  title: TournamentInsights Intelligence API
  description: Data intelligence and partner monetization tools for TournamentInsights / RefereeInsights
  version: 1.3.0

servers:
  - url: https://partnermcp-git-main-rod-davis-projects.vercel.app

x-spec-url: https://partnermcp-git-main-rod-davis-projects.vercel.app/api/openapi

security:
  - bearerAuth: []

components:
  schemas: {}
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

paths:
  /api/action:
    post:
      operationId: invokeTool
      summary: Invoke a TournamentInsights data or partner tool
      description: >
        Pass "tool" plus any tool-specific parameters at the top level of the
        request body. Do not nest parameters. Include parameters alongside "tool".
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tool
              additionalProperties: true
              properties:
                tool:
                  type: string
                  description: Name of the tool to invoke.
                  enum:
                    - get_dataset_health
                    - get_summary_dashboard
                    - get_trends
                    - get_state_sport_coverage
                    - get_missing_fields
                    - get_organizer_clusters
                    - get_venue_clusters
                    - get_missing_venues
                    - get_tournament_venue_worklist
                    - get_association_dashboard
                    - get_email_outreach_numbers
                    - get_top_organizer_domains
                    - get_tournaments_by_domain
                    - get_unverified_tournaments
                    - get_tournaments_missing_source_urls
                    - get_admin_reference
                    - get_partner_pipeline
                    - get_partner_links
                    - get_partner_click_summary
                    - get_fanatics_routing
                    - get_partner_knowledge
                    - mcp_healthcheck
                    - insert_research_note
                    - insert_tournament_candidate
                    - upsert_organizer_watchlist
                    - upsert_partner
                    - upsert_partner_link
                    - insert_partner_note
                    - update_partner_status
                    - upsert_partner_placement
                    - insert_partner_test_result
                    - get_roll_forward_log
                    - upsert_roll_forward_log
                    - list_tools
                    - get_search_runs
                    - get_search_run_findings
                    - get_search_coverage
                    - get_next_search_priorities
                    - insert_tournament_search_run
                    - insert_tournament_search_scope
                    - insert_tournament_search_finding
                    - insert_tournament_search_findings
                    - finalize_tournament_search_run
                    - get_search_organizer_intelligence
                    - insert_search_organizer_intelligence

                sport:
                  type: string
                  description: Single sport filter.
                  enum:
                    - soccer
                    - baseball
                    - softball
                    - lacrosse
                    - basketball
                    - hockey
                    - volleyball
                    - futsal

                state:
                  type: string
                  description: Single 2-letter US state code, for example CA or TX.

                sports:
                  type: array
                  items:
                    type: string
                  description: Multiple sport filter.

                states:
                  type: array
                  items:
                    type: string
                  description: Multiple 2-letter US state codes.

                limit:
                  type: integer
                  description: Maximum rows to return.

                offset:
                  type: integer
                  description: Pagination offset, 0-based.

                venue_status:
                  type: string
                  enum:
                    - missing
                    - incomplete
                    - missing_or_incomplete
                    - complete
                    - any
                  description: Venue completeness filter.

                date_from:
                  type: string
                  description: Start of date window, YYYY-MM-DD.

                date_to:
                  type: string
                  description: End of date window, YYYY-MM-DD.

                weeks_from_now_start:
                  type: integer
                  description: Weeks from today to start the date window.

                weeks_from_now_end:
                  type: integer
                  description: Weeks from today to end the date window.

                min_tournaments:
                  type: integer
                  description: Minimum tournament count to include.

                missing_any_of:
                  type: array
                  items:
                    type: string
                    enum:
                      - official_website_url
                      - tournament_director
                      - tournament_director_email
                      - referee_contact
                      - referee_contact_email
                      - start_date
                      - end_date
                      - city
                      - state
                  description: Fields to check for missingness.

                domain:
                  type: string
                  description: Root domain to filter by, for example gotsoccer.com.

                period:
                  type: string
                  enum:
                    - week
                    - month
                    - quarter
                  description: Bucketing period for trends.

                lookback_periods:
                  type: integer
                  description: Number of periods to return.

                query:
                  type: string
                  description: Search query.

                section:
                  type: string
                  description: Named knowledge/reference section to retrieve.

                partner_id:
                  type: string
                  description: Partner ID.

                start_date:
                  type: string
                  description: Start date, YYYY-MM-DD.

                end_date:
                  type: string
                  description: End date, YYYY-MM-DD.

                key:
                  type: string
                  description: Partner key.

                name:
                  type: string
                  description: Display name.

                status:
                  type: string
                  description: Status value.

                organizer_domain:
                  type: string
                  description: Domain to watchlist.

                reason:
                  type: string
                  description: Reason text.

                tags:
                  type: array
                  items:
                    type: string
                  description: Tags.

                source_url:
                  type: string
                  description: Source URL.

                entity_type:
                  type: string
                  description: Entity type.

                entity_id:
                  type: string
                  description: Entity ID.

                title:
                  type: string
                  description: Title.

                note:
                  type: string
                  description: Note text.

                partner_key:
                  type: string
                  description: Partner key.

                note_type:
                  type: string
                  description: Note type.

                label:
                  type: string
                  description: Display label.

                url:
                  type: string
                  description: Affiliate or target URL.

                placement_type:
                  type: string
                  description: Placement type.

                experiment_name:
                  type: string
                  description: Experiment name.

                ctr:
                  type: number
                  description: Click-through rate.

                clicks:
                  type: integer
                  description: Click count.

                conversions:
                  type: integer
                  description: Conversion count.

                notes:
                  type: string
                  description: Freeform notes.

                tested_at:
                  type: string
                  description: Test date, YYYY-MM-DD.

                region_name:
                  type: string
                  description: Free-text region label for a search run, for example "Southern California".

                search_run_id:
                  type: string
                  description: ID of a tournament_search_runs row.

                search_scope_id:
                  type: string
                  description: ID of a tournament_search_run_scopes row.

                supersedes_finding_id:
                  type: string
                  description: ID of the prior finding this one replaces (marks the prior finding non-current).

                candidate_status:
                  type: string
                  enum:
                    - Qualified
                    - Needs Venue Verification
                    - Needs Address Verification
                    - Needs Date Verification
                    - Out of Scope - State
                    - Out of Scope - Date
                    - Out of Scope - Sport
                    - Not a Tournament
                    - Duplicate
                    - Other Exclusion
                  description: Tournament-discovery finding status.

                tournament_name:
                  type: string
                  description: Candidate tournament name for a search-history finding.

                venue_name:
                  type: string
                  description: Venue name for a search-history finding.

                venue_address:
                  type: string
                  description: Venue street address for a search-history finding.

                venue_city:
                  type: string
                  description: Venue city for a search-history finding.

                venue_state:
                  type: string
                  description: Venue 2-letter state code for a search-history finding.

                venue_source_url:
                  type: string
                  description: Source URL verifying the venue.

                existing_tournament_id:
                  type: string
                  description: Production tournaments.id this finding may correspond to (reference only; never auto-linked).

                organizer_name:
                  type: string
                  description: Organizer display name for a search-history finding.

                search_prompt_version:
                  type: string
                  description: Version label of the research prompt used for a search run.

                search_prompt_text:
                  type: string
                  description: Full research prompt text for a search run (hashed and truncated in storage if over ~64KB).

                search_prompt_hash:
                  type: string
                  description: Precomputed hash of the research prompt, if known.

                search_method:
                  type: string
                  description: Free-text description of the search methodology, for example "venue-first deep geo discovery".

                research_agent:
                  type: string
                  description: Name/identifier of the agent that performed the search.

                research_model:
                  type: string
                  description: Model identifier used for the search.

                searched_at:
                  type: string
                  description: Timestamp the search was performed, ISO 8601.

                completed_at:
                  type: string
                  description: Timestamp the search run was completed/finalized, ISO 8601.

                searched_by:
                  type: string
                  description: Identifier of who/what ran the search.

                candidates_found:
                  type: integer
                  description: Total candidates found (run- or scope-level metric).

                qualified_rows:
                  type: integer
                  description: Count of Qualified findings (run- or scope-level metric).

                needs_venue_verification:
                  type: integer
                  description: Count of findings needing venue verification.

                needs_address_verification:
                  type: integer
                  description: Count of findings needing address verification.

                needs_date_verification:
                  type: integer
                  description: Count of findings needing date verification.

                duplicates_found:
                  type: integer
                  description: Count of findings marked Duplicate.

                out_of_scope_found:
                  type: integer
                  description: Count of out-of-scope/not-a-tournament/other-exclusion findings.

                organizer_domains:
                  type: array
                  items:
                    type: string
                  description: Organizer domains discovered during a search run.

                organizer_names:
                  type: array
                  items:
                    type: string
                  description: Organizer display names discovered during a search run.

                venue_names:
                  type: array
                  items:
                    type: string
                  description: Venue names discovered during a search run.

                high_value_sources:
                  type: array
                  items:
                    type: string
                  description: High-value monitoring source URLs discovered during a search run.

                search_summary:
                  type: string
                  description: Freeform summary of a search run.

                unresolved_work:
                  type: string
                  description: Freeform description of unresolved follow-up work for a search run.

                next_action:
                  type: string
                  description: Freeform description of the recommended next action for a search run.

                next_search_after:
                  type: string
                  description: Date (YYYY-MM-DD) after which this state/sport should be searched again.

                confidence_level:
                  type: string
                  enum:
                    - High
                    - Medium
                    - Low
                  description: Evidence-based confidence level for organizer intelligence.

                evidence_summary:
                  type: string
                  description: Summary of the evidence supporting the organizer intelligence. Required for all confidence levels.

                organizer_name:
                  type: string
                  description: Display name of the organizer (e.g. "NCS Southwest Arizona / Bruno Cafarelli").

                tournament_families:
                  type: array
                  items:
                    type: string
                  description: Recurring tournament name families attributed to this organizer. Must be explicitly evidenced.

                venue_clusters:
                  type: array
                  items:
                    type: string
                  description: Facilities this organizer recurringly uses. Not inferred from a single-event venue association.

                monitoring_urls:
                  type: array
                  items:
                    type: string
                  description: HTTP/HTTPS monitoring URLs for the organizer. Invalid URL rejects the entire array.

                recommended_cadence:
                  type: string
                  description: "Free-text monitoring cadence. Examples: Weekly, Monthly, Weekly; daily during the two weeks before each event."

                registration_platform:
                  type: string
                  description: Registration platform used by the organizer (descriptive only; not a proxy for organizer identity).

                scheduling_platform:
                  type: string
                  description: Scheduling/results platform used by the organizer (descriptive only).

                next_monitor_from:
                  type: string
                  description: Lower bound for next_monitor_after filter, YYYY-MM-DD.

                next_monitor_to:
                  type: string
                  description: Upper bound for next_monitor_after filter, YYYY-MM-DD.

                target_year:
                  type: integer
                  description: Target year (2020–2040) for roll-forward log filtering or upsert.

                batch_label:
                  type: string
                  description: Batch label for grouping roll-forward log entries.

                parent_tournament_id:
                  type: string
                  description: UUID of the parent tournament for a roll-forward log entry.

                sibling_id:
                  type: string
                  description: UUID of the sibling tournament found for this target year.

                researched_at:
                  type: string
                  description: ISO 8601 timestamp when the roll-forward entry was researched.

                source_batch_id:
                  type: string
                  description: Caller-supplied idempotency key for insert_tournament_search_run.

                window_from:
                  type: string
                  description: Lower bound of a tournament date window (YYYY-MM-DD) for search-history read tools.

                window_to:
                  type: string
                  description: Upper bound of a tournament date window (YYYY-MM-DD) for search-history read tools.

                searched_from:
                  type: string
                  description: Lower bound on searched_at (ISO 8601) for search-history read tools.

                searched_to:
                  type: string
                  description: Upper bound on searched_at (ISO 8601) for search-history read tools.

                current_only:
                  type: boolean
                  description: For get_search_run_findings — restrict to current (non-superseded) findings. Defaults to true.

                findings:
                  type: array
                  description: For insert_tournament_search_findings — array of finding objects (same shape as insert_tournament_search_finding params, omitting search_run_id).
                  items:
                    type: object
                    additionalProperties: true

      responses:
        "200":
          description: Successful tool result.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                properties:
                  result:
                    type: object
                    description: Tool output. Structure varies by tool.
                    additionalProperties: true

        "400":
          description: Unknown tool name or invalid request.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                properties:
                  error:
                    type: string

        "401":
          description: Missing or invalid bearer token.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                properties:
                  error:
                    type: string

        "500":
          description: Tool execution error.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                properties:
                  error:
                    type: string